DAH-1160 Add lottery results and application due date to mobile, upgrade minor version of node
