+++ app/javascript/modules/listingDetailsAside/ListingDetailsAside.tsx
@@ -6,7 +6,7 @@ import { ListingDetailsProcess } from "./ListingDetailsProcess"
 import { isOpen } from "../../util/listingUtil"
 import { ListingDetailsApply } from "./ListingDetailsApply"
 import { ListingDetailsApplicationDate } from "./ListingDetailsApplicationDate"
-import { ListingDetailsLottery } from "../listingDetailsLottery/ListingDetailsLottery"
+import { ListingDetailsLotteryResults } from "../listingDetailsLottery/ListingDetailsLotteryResults"
 
 export interface ListingDetailsSidebarProps {
   listing: RailsListing
@@ -28,7 +28,7 @@ export const ListingDetailsAside = ({ listing, imageSrc }: ListingDetailsSidebar
       <aside className="w-full static md:absolute md:right-0 md:w-1/3 md:top-0 sm:w-2/3 md:ml-2 h-full md:border border-solid bg-white">
         <div className="hidden md:block">
           <ListingDetailsApplicationDate isApplicationOpen={isApplicationOpen} listing={listing} />
-          <ListingDetailsLottery listing={listing} />
+          <ListingDetailsLotteryResults listing={listing} />
           {isApplicationOpen && <ListingDetailsInfoSession listing={listing} />}
           <ListingDetailsApply listing={listing} />
           <ListingDetailsProcess listing={listing} />
