+++ app/javascript/modules/listingDetailsLottery/ListingDetailsLottery.tsx
@@ -1,81 +1,26 @@
-import React, { useEffect, useState } from "react"
-import {
-  AppearanceSizeType,
-  AppearanceStyleType,
-  Button,
-  Heading,
-  Modal,
-  t,
-} from "@bloom-housing/ui-components"
+import React from "react"
 import { RailsListing } from "../listings/SharedHelpers"
+import { ListingDetailItem, t } from "@bloom-housing/ui-components"
+import { ListingDetailsLotteryResults } from "./ListingDetailsLotteryResults"
 import { isLotteryComplete } from "../../util/listingUtil"
-import { getLotteryBucketDetails } from "../../api/listingApiService"
-import { RailsLotteryResult } from "../../api/types/rails/listings/RailsLotteryResult"
-import { ListingDetailsLotterySearchForm } from "./ListingDetailsLotterySearchForm"
-import { localizedFormat, renderInlineWithInnerHTML } from "../../util/languageUtil"
-import ErrorBoundary, { BoundaryScope } from "../../components/ErrorBoundary"
 
 export interface ListingDetailsLotteryProps {
+  imageSrc: string
   listing: RailsListing
 }
 
-export const ListingDetailsLottery = ({ listing }: ListingDetailsLotteryProps) => {
-  const [lotteryBucketDetails, setLotteryBucketDetails] = useState<RailsLotteryResult>()
-  const [isModalOpen, setIsModalOpen] = useState(false)
-
-  useEffect(() => {
-    if (isLotteryComplete(listing)) {
-      void getLotteryBucketDetails(listing.Id).then((lotteryBucketDetails) => {
-        setLotteryBucketDetails(lotteryBucketDetails)
-      })
-    }
-  }, [listing, listing.Id])
-
+export const ListingDetailsLottery = ({ imageSrc, listing }: ListingDetailsLotteryProps) => {
   return (
+    listing &&
     isLotteryComplete(listing) && (
-      <div className="border-b pt-4 text-center">
-        <ErrorBoundary boundaryScope={BoundaryScope.component}>
-          <Heading className="mb-4" priority={4}>
-            {t("lottery.lotteryResults")}
-          </Heading>
-          <p className="mb-4 text-sm uppercase">
-            {localizedFormat(listing.Lottery_Results_Date, "LL")}
-          </p>
-          <div className="bg-gray-100 py-4">
-            {listing.Lottery_Summary && (
-              <div className="mb-3 mx-2 text-gray-700 text-sm">
-                {renderInlineWithInnerHTML(listing.Lottery_Summary)}
-              </div>
-            )}
-            <Button
-              size={AppearanceSizeType.small}
-              styleType={AppearanceStyleType.primary}
-              onClick={() => {
-                setIsModalOpen(true)
-              }}
-            >
-              {t("listings.lottery.viewLotteryResults")}
-            </Button>
-          </div>
-
-          {lotteryBucketDetails && (
-            <Modal
-              onClose={() => setIsModalOpen(false)}
-              open={isModalOpen}
-              title=""
-              modalClassNames="max-w-0"
-              innerClassNames="p-0"
-              closeClassNames="z-50"
-              scrollable
-            >
-              <ListingDetailsLotterySearchForm
-                listing={listing}
-                lotteryBucketDetails={lotteryBucketDetails}
-              />
-            </Modal>
-          )}
-        </ErrorBoundary>
-      </div>
+      <ListingDetailItem
+        imageAlt={""}
+        imageSrc={imageSrc}
+        title={t("lottery")}
+        subtitle={t("lottery.lotteryInfoSubheader")}
+      >
+        <ListingDetailsLotteryResults listing={listing} />
+      </ListingDetailItem>
     )
   )
 }
