+++ app/javascript/modules/listingDetailsLottery/ListingDetailsLotterySearchForm.tsx
@@ -100,7 +100,7 @@ export const ListingDetailsLotterySearchForm = ({
     (lotteryFormStatus === LOTTERY_SEARCH_FORM_STATUS.API_ERROR && t("error.lotteryRankingSearch"))
 
   return (
-    <div>
+    <div className="lottery-search-form">
       <header className="pt-8 pb-4 text-center bg-white sticky top-0 z-50">
         <Heading>{t("lottery.lotteryResults")}</Heading>
         <h2 className="font-sans font-semibold text-sm uppercase">{listing.Name}</h2>
